+++ Traits/XMLFixtureLoaderTrait.php
@@ -1,0 +1,34 @@
+<?php
+/*
+ * This file is part of the Sulu CMS.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace Sulu\Bundle\ProductBundle\Traits;
+
+trait XMLFixtureLoaderTrait
+{
+    /**
+     * @param string $xmlFile
+     * @param string $nodePath
+     *
+     * @return \DOMNodeList
+     */
+    protected function loadElementsFromXmlFileWithPath($xmlFile, $nodePath)
+    {
+        $xml = file_get_contents($xmlFile);
+        $xml = preg_replace('/>\s+</', '><', $xml);
+
+        $doc = new \DOMDocument();
+        $doc->loadXML($xml);
+
+        $xpath = new \DOMXpath($doc);
+        $elements = $xpath->query($nodePath);
+
+        return $elements;
+    }
+}