+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu Product Bundle
 =================================
 
 * dev-develop
+    * FEATURE Added key to attributes
     * BUGFIX Fixed tests of special prices
 
 * 0.11.0 (2016-07-06)