+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 CHANGELOG for Sulu Product Bundle
 =================================
 
-* dev-develop
+* 0.11.1 (2016-07-11)
+
     * FEATURE   Added possibility to add Attribute fixtures by adding the file name to config.yml
     * FEATURE   Added key to attributes
     * BUGFIX    Fixed tests of special prices
 
 * 0.11.0 (2016-07-06)
+
     * ENHANCEMENT Added methods for price calculations of addons
     * ENHANCEMENT Added flag for recurring price
     * ENHANCEMENT Added API for product addons
@@ -14,21 +16,26 @@ CHANGELOG for Sulu Product Bundle
     * ENHANCEMENT Added UI for addons.
 
 * 0.10.7 (2016-07-04)
+
     * BUGFIX Rendering prices tab with special prices in admin
 
 * 0.10.6 (2016-06-23)
+
     * BUGFIX Fixed issue with tags when adding a new product
     * ENHANCEMENT Added tests for tags in products
 
 * 0.10.5 (2016-06-23)
+
     * FEATURE Added tags to product
     * ENHANCEMENT Improved ProductTestData by adding product statuses
     
 * 0.10.4 (2016-06-21)
+
     * ENHANCEMENT Improved ContactTestData by adding use of ContactRepositories create contact function
     * ENHANCEMENT Cleanups in ProductPriceManager
 
 * 0.10.3 (2016-06-15)
+
     * ENHANCEMENT Removed not used attribute types from fixtures
     * BUGFIX Display fallback locale when adding a new attribute value to a product
     * BUGFIX Generating thumbnails